refactor(types): use types-ambient naming
