+++ packages/boot/index.js
@@ -1,3 +1,3 @@
 // @ts-check
 // Ambient types
-import '@agoric/vats/src/core/types.js';
+import '@agoric/vats/src/core/types-ambient.js';
