+++ packages/inter-protocol/test/smartWallet/test-oracle-integration.js
@@ -1,4 +1,4 @@
-import '@agoric/vats/src/core/types.js';
+import '@agoric/vats/src/core/types-ambient.js';
 
 import { test as anyTest } from '@agoric/zoe/tools/prepare-test-env-ava.js';
