+++ packages/zoe/src/contractFacet/zcfMint.js
@@ -10,7 +10,7 @@ import { addToAllocation, subtractFromAllocation } from './allocationMath.js';
 import '../internal-types.js';
 import { ZcfMintI } from '../typeGuards.js';
 import './internal-types.js';
-import './types.js';
+import './types-ambient.js';
 
 const { Fail } = assert;
